--- src/actions/documentActions.js
@@ -1,5 +1,0 @@
-export const UPDATE_MARKDOWN = 'UPDATE_MARKDOWN';
-export const updateMarkdown = payload => ({
-  type: UPDATE_MARKDOWN,
-  payload: payload
-});
