+++ src/components/navigation/Tabs.js
@@ -2,13 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const Tabs = ({ files, handleTabSelect }) => {
-  const tabElements = Object.values(files).map(file => {
+  const tabElements = Object.keys(files).map(fileName => {
     return (
       <li 
-        key={file.id || file.fileName} 
-        name={file.fileName} 
-        onClick={() => handleTabSelect(file.fileName)}>
-        {file.fileName}
+        key={fileName} 
+        name={fileName} 
+        onClick={() => handleTabSelect(fileName)}>
+        {fileName}
       </li>
     );
   });
