--- src/reducers/documentReducer.js
@@ -1,10 +1,0 @@
-// export const initialState = {
-//   markdown: '# Hi there'
-// };
-
-// export default function reducer(state = initialState, action) {
-//   switch(action.type) {
-//     default:
-//       return state;
-//   }
-// }
