+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from 'redux';
-import documents from './documentReducer';
 import navigation from './navigationReducer';
 
 export default combineReducers({
-  documents,
   navigation
 });
