--- src/selectors/documentSelectors.js
@@ -1,1 +1,0 @@
-export const getMarkdown = state => state.navigation.files[state.navigation.activeDocument].markdown;
