+++ src/selectors/navigationSelectors.js
@@ -1,3 +1,4 @@
 export const getFiles = state => state.navigation.files;
 export const getNewFileName = state => state.navigation.newFileName;
 export const getUpdatedMarkdown = state => state.documents.markdown;
+export const getMarkdown = state => state.navigation.files[state.navigation.activeDocument].markdown;
